fix(script): create_model

+++ scripts/create_model.zsh
@@ -8,13 +8,16 @@ IN_FILE="output/unknown_cmd/${UPPER_SNAKE}.json"
 OUT_FILE="src/blivedm/models/${LOWER_SNAKE}.py"
 TMP_NAME="_tmp_create_model"
 
-[ -f "$IN_FILE" ] || echo "$IN_FILE not exists" && exit 2
+if [ -f "$IN_FILE" ]; then
+  echo "$IN_FILE not exists"
+  exit 2
+fi
 
-jq -s '.[0]' "$IN_FILE" > ${TMP_NAME}.json
+jq -s '.[0]' "$IN_FILE" >${TMP_NAME}.json
 
 datamodel-codegen --input ${TMP_NAME}.json --output ${TMP_NAME}.py --input-file-type json
 
-cat >"$OUT_FILE" << EOF
+cat >"$OUT_FILE" <<EOF
 from ._base import *
 
 
@@ -24,11 +27,9 @@ $(cat ${TMP_NAME}.py | head -n-8 | tail -n+10)
 class ${CAPITAL_CASE}Command(CommandModel):
     cmd: Literal['${UPPER_SNAKE}']
     data: Data
-
 EOF
 
 rm -f "${TMP_NAME}.json" "${TMP_NAME}.py"
 
 echo "lines should be checked:"
 grep -E "time|color" "${OUT_FILE}"
-
